+++ R/filteredDownloads.R
@@ -25,19 +25,24 @@ filteredDownloads <- function(packages = "HistData", date = NULL,
 
   ymd <- logDate(date)
   cran_log <- packageLog(packages = packages, date = ymd,
-    memoization = memoization, multi.core = cores)
+    memoization = memoization, check.package = FALSE, multi.core = cores)
 
   if (is.data.frame(cran_log)) {
     ct <- nrow(cran_log)
   } else {
     ct <- vapply(cran_log, nrow, integer(1L))
   }
+  
+  individual.filter <- (ip.filter | triplet.filter | small.filter | 
+                        sequence.filter | size.filter) & all.filters
+  
+  if (individual.filter) all.filters <- FALSE
 
   f.cran_log <- packageLog(packages = packages, date = ymd,
     all.filters = all.filters, ip.filter = ip.filter,
     triplet.filter = triplet.filter, small.filter = small.filter,
     sequence.filter = sequence.filter, size.filter = size.filter,
-    memoization = memoization, multi.core = cores)
+    memoization = memoization, check.package = FALSE, multi.core = cores)
 
   if (is.data.frame(f.cran_log)) {
     f.ct <- nrow(f.cran_log)